# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API untuk sistem service motor. Sebagian besar endpoint membutuhkan token JWT (Bearer) yang diperoleh dari login/register.",
        "title": "Sistem Service Motor API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/vehicle-taxes": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil daftar pajak kendaraan milik user yang login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vehicle Tax"
                ],
                "summary": "Ambil semua pajak kendaraan user",
                "responses": {
                    "200": {
                        "description": "Data pajak kendaraan berhasil diambil",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "token tidak valid",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Membuat data pajak kendaraan baru untuk user yang login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vehicle Tax"
                ],
                "summary": "Buat pajak kendaraan",
                "parameters": [
                    {
                        "description": "Data pajak kendaraan",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/example_com_my-api_internal_dto.CreateVehicleTaxRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Data pajak kendaraan berhasil dibuat",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "request body tidak valid",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "token tidak valid",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "kendaraan tidak ditemukan",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/vehicle-taxes/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil data pajak kendaraan milik user yang login berdasarkan ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vehicle Tax"
                ],
                "summary": "Ambil pajak kendaraan berdasarkan ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID pajak",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data pajak kendaraan berhasil diambil",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "id tidak valid",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "token tidak valid",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "pajak kendaraan tidak ditemukan",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Memperbarui data pajak kendaraan milik user yang login berdasarkan ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vehicle Tax"
                ],
                "summary": "Update pajak kendaraan",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID pajak",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Data pajak kendaraan",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/example_com_my-api_internal_dto.UpdateVehicleTaxRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data pajak kendaraan berhasil diupdate",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "request body tidak valid",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "token tidak valid",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "pajak kendaraan tidak ditemukan",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Menghapus data pajak kendaraan milik user yang login berdasarkan ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vehicle Tax"
                ],
                "summary": "Hapus pajak kendaraan",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID pajak",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "berhasil hapus data pajak kendaraan",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "id tidak valid",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "token tidak valid",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "pajak kendaraan tidak ditemukan",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "example_com_my-api_internal_dto.CreateVehicleTaxRequest": {
            "type": "object",
            "required": [
                "due_date",
                "other_amount",
                "payment_date",
                "pkb_amount",
                "swdkllj_amount",
                "tax_year",
                "vehicle_id"
            ],
            "properties": {
                "due_date": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "other_amount": {
                    "type": "number",
                    "minimum": 0
                },
                "payment_date": {
                    "type": "string"
                },
                "pkb_amount": {
                    "type": "number",
                    "minimum": 0
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "paid",
                        "unpaid"
                    ]
                },
                "swdkllj_amount": {
                    "type": "number",
                    "minimum": 0
                },
                "tax_year": {
                    "type": "integer",
                    "maximum": 2100
                },
                "vehicle_id": {
                    "type": "integer"
                }
            }
        },
...
        "example_com_my-api_internal_dto.UpdateVehicleTaxRequest": {
            "type": "object",
            "required": [
                "due_date",
                "other_amount",
                "payment_date",
                "pkb_amount",
                "swdkllj_amount",
                "tax_year",
                "vehicle_id"
            ],
            "properties": {
                "due_date": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "other_amount": {
                    "type": "number",
                    "minimum": 0
                },
                "payment_date": {
                    "type": "string"
                },
                "pkb_amount": {
                    "type": "number",
                    "minimum": 0
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "paid",
                        "unpaid"
                    ]
                },
                "swdkllj_amount": {
                    "type": "number",
                    "minimum": 0
                },
                "tax_year": {
                    "type": "integer",
                    "maximum": 2100
                },
                "vehicle_id": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Masukkan token JWT dengan format: `Bearer \u003ctoken\u003e`",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}